pyproject.toml
[project]
name = "vit-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "pymongo>=4.14.1",
    "redis>=6.4.0",
]
